+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/entity/listing/CertifiedProductEntity.java
@@ -346,7 +346,7 @@ public class CertifiedProductEntity {
     protected Date creationDate;
 
     @Basic(optional = false)
-    @Column(nullable = false)
+    @Column(nullable = false, updatable = false)
     protected Boolean deleted;
 
     @Basic(optional = false)